Fix issues caused by dist-upgrade
Using a dist-upgrade instead of an upgrade meant that the kernel
version being patched by the VirtualBox Guest Additions was not the
one that ended up running in the final box. To rectify this, we need to
reboot after the dist-upgrade and before the rest of the stuff happens.
I also added in code to remove old kernel versions and to clean up the
apt cache, which reduced the size of the final box file by about 200 MB.